+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Support rails_event_store 3.0 by using the renamed `RailsEventStore::AfterCommitDispatcher`, `RubyEventStore::SyncScheduler` and `RubyEventStore::ActiveRecord::EventRepository`, falling back to the pre-2.19 names when running on older versions. ([#24][])
+
 ## 1.3.0 (2025-12-04)
 
 - Update Rails Event Store configuration to avoid deprecation warning ([@christophermlne][], [@phil-monroe][])
@@ -51,3 +53,4 @@ Now `ActiveSupport.on_load(:active_event_store) { ... }` works.
 [@chriscz]: https://github.com/chriscz
 [@caws]: https://github.com/caws
 [@Samsinite]: https://github.com/Samsinite
+[#24]: https://github.com/palkan/active_event_store/issues/24